Move check_dist_requires_python() to resolve.py.
