+++ src/pip/_internal/utils/packaging.py
@@ -6,7 +6,6 @@ from email.parser import FeedParser
 from pip._vendor import pkg_resources
 from pip._vendor.packaging import specifiers, version
 
-from pip._internal import exceptions
 from pip._internal.utils.misc import display_path
 from pip._internal.utils.typing import MYPY_CHECK_RUNNING
 
@@ -56,31 +55,6 @@ def get_metadata(dist):
     feed_parser = FeedParser()
     feed_parser.feed(metadata)
     return feed_parser.close()
-
-
-def check_dist_requires_python(dist, version_info):
-    """
-    :param version_info: A 3-tuple of ints representing the Python
-        major-minor-micro version to check (e.g. `sys.version_info[:3]`).
-    """
-    pkg_info_dict = get_metadata(dist)
-    requires_python = pkg_info_dict.get('Requires-Python')
-    try:
-        if not check_requires_python(
-            requires_python, version_info=version_info,
-        ):
-            raise exceptions.UnsupportedPythonVersion(
-                "%s requires Python '%s' but the running Python is %s" % (
-                    dist.project_name,
-                    requires_python,
-                    '.'.join(map(str, version_info)),)
-            )
-    except specifiers.InvalidSpecifier as e:
-        logger.warning(
-            "Package %s has an invalid Requires-Python entry %s - %s",
-            dist.project_name, requires_python, e,
-        )
-        return
 
 
 def get_installer(dist):
